Added "build" and "run" commands
I removed initialising git repository to project generation because I felt the output was annoying, I will aim to use libgit2 to generate it instead which should not produce an output (or maybe there's a git flag to minimise console output).
Building and running a project are now done!
Building will generate an executable in the /src/ folder.
Running will also generate the same executable but run it as well.