Added new_project() error handling and cli::question()
New error handling should handle every error case and response with either an attempt to recover or abort message that is specific to the scenario.
This means if, during project creation, a lack of permissions is found the abort message will tell the user.
If a recoverable issues such as the project directory already exists, then the user will be questions and action will be taken.

This update also adds a new cli associated function: cli::question(msg) for yes/no questions.

Errors handled by cli::new_project():
- NotFound
- PermissionDenied
- AlreadyExists
- OutOfMemory
- _ (generic fallback)

This should be more user-friendly, and other cli functions will be amended as well.