Added error handling for run_project() and build_project()
Error handling is important for the user experience.
Getting a really confusing error where the user has no idea what's going on, or how to fix it makes the user experience significantly worse.
In order to fix this, for the last few commits, I've been adding error handling that uses the error type (error.kind()) to find the specific issue and report the issue to the user.

In this commit, errors that bubble up to run_project() and build_project() have now been handled.
If successful (no errors), a simple "[SUCCESS] {message}" is displayed.
Otherwise, a custom abort message is sent to the user with a brief explanation of what happened.

Changelog:
- Added child process wait method to allow command execution to finish before continuing with the main thread in manager.rs
- Added generate_executable_error_handler() to handle errors for run_project() and build_project()