Config.toml deserialization complete
I've set up the correct structures needed for the config.toml to be serialized and deserialized.
Three new structures have been added to do this:
1. Config struct: stores everything inside the config (project, run profile, build profile)
2. Project struct: stores everything inside the [project] table
3. Profile sturct: stores everything inside the compilation tables: [profile-run], [profile-build]

Originally, as mentioned in the issue, the usage of dotted tables and dotted keys was heavy.
However, some issues with toml-rs has caused these to be wrapped in double quotes.
For now, config.toml will use dashes and underscores.

Changelog:
- Added Project struct
- Added Profile struct