Added serialisation and fixed profile naming bug
Re-added Config::load() which can be used to convert a project's config.toml into a Config struct.
This will allow rectx to read the config file and run accordingly.
So, for example, if the user were to change the source_main from "main.rct" to "epic_new_main.rct" the build command would still understand how to build the project because the main file used can now be from the config.toml file and not from a built-in program constant.

I've also fixed a bug to do with serialisation namingfor profile build and profile run.