Fixed command stream halt if no new line
This is a weird bug I found but if the commands file doesn't end with a new line then the last command won't be read.
The current fix will check the byte buffer (used for reading) and if it's larger than 0 but not a new line then it will add a final command.